Fix 404 error when clicking author avatar in video reels
🐛 BUG FIX:

Fixed the 404 error that occurred when clicking on author avatars
in the video reels component.

Issue:
- Clicking author avatar tried to navigate to /student/profile/{authorId}
- This route doesn't exist, causing a 404 error
- The student profile page is at /student/profile without user ID parameter

Solution:
- Removed the clickable navigation from author avatars
- Changed from <button> to <div> to prevent navigation
- Author avatars now display as non-interactive elements
- This prevents the 404 error while maintaining visual design

The author information is still displayed but clicking it no longer
causes navigation errors. In the future, if we want to implement
viewing other students' profiles, we would need to create a proper
route like /student/profile/view/{userId} or similar.